requirements.txt
dash==2.13.0
dash-bootstrap-components==1.5.0
pandas==1.5.3
plotly==5.9.0
pandas_ta==0.3.14b0
google-cloud-storage==2.12.0